wip frontend preference flow, refactoring components
