+++ frontend/movie-recs/src/app/movies/[id]/page.tsx
@@ -2,13 +2,15 @@
 
 import MovieCard from "@/components/MovieCard";
 
+// TODO: add logic on button to add to favorite/liked movie list and save to context (if logged in, will write to db)
+
 const Movie = () => {
-    return (
-        <div>
-            <div className="text-lg">Movie</div>
-            <MovieCard/>
-        </div>
-    )
-}
+  return (
+    <div className="flex flex-col justify-center items-center content-center">
+      <div className="text-lg">Movie Details</div>
+      {/* TODO add movie details */}
+    </div>
+  );
+};
 
 export default Movie;
