+++ frontend/movie-recs/src/app/preferences/layout.tsx
@@ -1,9 +1,9 @@
 
 export default function PreferencesLayout({ children }: { children: React.ReactNode }) {
     return (
-        <div className="flex flex-col justify-center items-center content-center">
+        <>
         {/* // <div> */}
             {children}
-        </div>
+        </>
     )
 }
